requirements.txt
requests
streamlit
pandas
plotly

python-dotenv

boto3 

langchain
langchain-community
langchain-aws

scipy